server : merge oai.hpp in utils.hpp
